[#90465176] Add spec using selenium
Created a page object to drive the New Car data entry
Created a spec which uses the new page object
Spec fails right now due to not finding the success message; the
element has no ID or other obvious way for it to be found.
Probably better to go to an rspec matcher for the validation in this
scenario.